+++ src/controller_device_send.h
@@ -45,7 +45,7 @@ extern "C" {
 #endif
 
 int device_send_lock(clixon_handle h, device_handle dh, int lock);
-int device_send_get_config(clixon_handle h, device_handle ch, int s);
+int device_send_get(clixon_handle h, device_handle ch, int s, int state, char *xpath);
 int device_send_get_schema_next(clixon_handle h, device_handle dh, int s, int *nr);
 int device_send_get_schema_list(clixon_handle h, device_handle dh, int s);
 int device_create_edit_config_diff(clixon_handle h, device_handle dh,